refactoring ManagerUserPage
