+++ pages/ProjectPage.py
@@ -1,5 +1,9 @@
 from pages.BasePage import BasePage
 from selenium.webdriver.common.by import By
+from config.TestData import TestData as TD
+from pages.DashboardPage import DashboardPageLocators, DashboardPage
+from pages.ManageUserPage import ManageUserPage
+from pages.NewItemPage import NewItemPageLocators
 
 
 class ProjectPageLocators:
@@ -11,7 +15,6 @@ class ProjectPageLocators:
     BUILD_NOW = (By.XPATH, "//a[@title='Build Now']//span[2]")
     BUILD_HISTORY_JOBS = (By.XPATH, '//tr[contains(@class, "build-row multi-line")]')
     BUILD_SUCCESS_JOBS = (By.XPATH, '//a[@class="build-status-link"]')
-    # BUILD_SUCCESS_LAST_JOB = (By.XPATH, '//td[@class="build-row-cell"][1]//img')
     BUILD_SUCCESS_LAST_JOB = (By.XPATH, '//td[@class="build-row-cell"][1]//a[@class="build-status-link"]')
 
 class ProjectPage(BasePage):
@@ -23,3 +26,32 @@ class ProjectPage(BasePage):
     def __init__(self, driver):
         super().__init__(driver)
 
+    def create_new_job(self):
+        """
+        create new job Freestyle project
+        :return:
+        """
+        driver = DashboardPage(self.driver)
+        driver.click(DashboardPageLocators.TEXT_NEW_ITEM)
+        name = ManageUserPage.CREATE_USER_JOB
+        driver.do_send_keys(NewItemPageLocators.ENTER_AN_ITEM_NAME, name)
+        driver.click(NewItemPageLocators.FREESTYLE_PROJECT)
+        driver.get_wait_is_clickable(NewItemPageLocators.OK_BUTTON)
+        driver.click(NewItemPageLocators.OK_BUTTON)
+        driver.get_wait(NewItemPageLocators.SAVE_BUTTON)
+        driver.click(NewItemPageLocators.SAVE_BUTTON)
+        return name
+
+    def delete_job(self, name):
+        """
+        delete job from list
+        :param name:
+        :return:
+        """
+        URL_JOB_FOR_DELETE = TD.BASE_URL + f'job/{name}/'
+
+        driver = ProjectPage(self.driver)
+        driver.go_to_page(URL_JOB_FOR_DELETE)
+        driver.click(ManageUserPage.JOB_DELETE_PROJECT)
+        driver.get_wait_for_alert()
+        driver.accept_alert()
