+++ tests/test_manage_user.py
@@ -1,64 +1,13 @@
 import pytest
 
 from config.TestData import TestData as TD
-from pages.DashboardPage import DashboardPage, DashboardPageLocators
 from pages.LoginPage import LoginPage
 from pages.ManageUserPage import ManageUserPage
 from pages.PeoplePage import PeoplePage, URLLocators
-from pages.NewItemPage import NewItemPage, NewItemPageLocators
 from pages.ProjectPage import ProjectPageLocators, ProjectPage
 
 
 class TestManageUserPage:
-
-    # def get_element_from_people_page(self, locator):
-    #     """
-    #     get element from people page
-    #     :param locator:
-    #     :return:
-    #     """
-    #     driver = PeoplePage(self.driver)
-    #     driver.go_to_page(URLLocators.URL_PEOPLE)
-    #     lst = driver.get_elements_text(locator)
-    #     return lst
-
-    # def login_with_default_credential(self):
-    #     """
-    #     login with default credential
-    #     :return:
-    #     """
-    #     driver = LoginPage(self.driver)
-    #     driver.login_jenkins(TD.LOGIN, TD.PASSWORD)
-
-    def create_new_job(self):
-        """
-        create new job Freestyle project
-        :return:
-        """
-        driver = DashboardPage(self.driver)
-        driver.click(DashboardPageLocators.TEXT_NEW_ITEM)
-        name = ManageUserPage.CREATE_USER_JOB
-        driver.do_send_keys(NewItemPageLocators.ENTER_AN_ITEM_NAME, name)
-        driver.click(NewItemPageLocators.FREESTYLE_PROJECT)
-        assert driver.is_enabled(NewItemPageLocators.OK_BUTTON)
-        driver.click(NewItemPageLocators.OK_BUTTON)
-        driver.get_wait(NewItemPageLocators.SAVE_BUTTON)
-        driver.click(NewItemPageLocators.SAVE_BUTTON)
-        return name
-
-    def delete_job(self, name):
-        """
-        delete job from list
-        :param name:
-        :return:
-        """
-        URL_JOB_FOR_DELETE = TD.BASE_URL + f'job/{name}/'
-
-        driver = ProjectPage(self.driver)
-        driver.go_to_page(URL_JOB_FOR_DELETE)
-        driver.click(ManageUserPage.JOB_DELETE_PROJECT)
-        driver.get_wait_for_alert()
-        driver.accept_alert()
 
     def test_create_user_valid_credential(self):
         """
@@ -108,7 +57,7 @@ class TestManageUserPage:
         veify that the job was started
         :return:
         """
-        name = self.create_new_job()
+        name = ProjectPage.create_new_job(self)
         URL_JOB = TD.BASE_URL + f'job/{name}/'
 
         driver = ProjectPage(self.driver)
@@ -118,7 +67,7 @@ class TestManageUserPage:
         driver.click(ProjectPageLocators.BUILD_SUCCESS_LAST_JOB)
 
         assert driver.get_element_text(ManageUserPage.STARTED_BY_USER) == ManageUserPage.USER_FULLNAME
-        self.delete_job(name)
+        ProjectPage.delete_job(self, name)
 
     def test_delete_new_user(self):
         """
@@ -176,13 +125,14 @@ class TestManageUserPage:
         verify, that user present on the page
         :return:
         """
+        driver = LoginPage(self.driver)
+        driver.login_with_default_credential()
         driver = ManageUserPage(self.driver)
         driver.go_to_page(ManageUserPage.URL_USER_MANAGE)
         driver.click_button_create_new_user()
         driver.fill_all_field_and_click_save(ManageUserPage.USER_NAME_UNDERSCORE, ManageUserPage.PASSWORD)
 
         assert driver.get_elements_text(ManageUserPage.USER_ID_UNDERSCORE)[0] == ManageUserPage.USER_NAME_UNDERSCORE
-        self.test_delete_new_user()
 
     @pytest.mark.skip
     def test_create_user_with_hyphen_name(self):
